Moved shouldRun check into DeviceStateHolder.
