+++ src/main/java/com/nutomic/syncthingandroid/syncthing/DeviceStateHolder.java
@@ -4,10 +4,16 @@ import android.annotation.TargetApi;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.net.ConnectivityManager;
 import android.net.wifi.WifiInfo;
 import android.net.wifi.WifiManager;
 import android.os.BatteryManager;
+import android.preference.PreferenceManager;
+import android.util.Log;
+
+import java.util.HashSet;
+import java.util.Set;
 
 /**
  * Holds information about the current wifi and charging state of the device.
@@ -16,6 +22,8 @@ import android.os.BatteryManager;
  * to {@link #update(android.content.Intent)}.
  */
 public class DeviceStateHolder extends BroadcastReceiver {
+
+    private static final String TAG = "DeviceStateHolder";
 
     /**
      * Intent extra containing a boolean saying whether wifi is connected or not.
@@ -91,7 +99,52 @@ public class DeviceStateHolder extends BroadcastReceiver {
         }
     }
 
-    public String getWifiSsid() {
+    private String getWifiSsid() {
         return mWifiSsid;
     }
+
+    public boolean shouldRun() {
+        if (SyncthingService.alwaysRunInBackground(mContext)) {
+            // Always run, ignoring wifi/charging state.
+            return true;
+        }
+        else {
+            SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(mContext);
+            // Check wifi/charging state against preferences and start if ok.
+            boolean prefStopMobileData = prefs.getBoolean(SyncthingService.PREF_SYNC_ONLY_WIFI, false);
+            boolean prefStopNotCharging = prefs.getBoolean(SyncthingService.PREF_SYNC_ONLY_CHARGING, false);
+
+            return (isCharging() || !prefStopNotCharging) &&
+                    (!prefStopMobileData || isAllowedWifiConnected());
+        }
+    }
+
+    private boolean isAllowedWifiConnected() {
+        boolean wifiConnected = isWifiConnected();
+        if (wifiConnected) {
+            SharedPreferences sp = PreferenceManager.getDefaultSharedPreferences(mContext);
+            Set<String> ssids = sp.getStringSet(SyncthingService.PREF_SYNC_ONLY_WIFI_SSIDS, new HashSet<String>());
+            if (ssids.isEmpty()) {
+                Log.d(TAG, "All SSIDs allowed for syncing");
+                return true;
+            } else {
+                String ssid = getWifiSsid();
+                if (ssid != null) {
+                    if (ssids.contains(ssid)) {
+                        Log.d(TAG, "SSID [" + ssid + "] found in whitelist: " + ssids);
+                        return true;
+                    }
+                    Log.i(TAG, "SSID [" + ssid + "] not whitelisted: " + ssids);
+                    return false;
+                } else {
+                    // Don't know the SSID (yet) (should not happen?!), so not allowing
+                    Log.w(TAG, "SSID unknown (yet), cannot check SSID whitelist. Disallowing sync.");
+                    return false;
+                }
+            }
+        }
+        Log.d(TAG, "Wifi not connected");
+        return false;
+    }
+
 }
